requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.0.2
django-heroku==0.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
boto3==1.35.39
django-storages==1.14.4